refactor(web): lift fetchAssistantIdentity + AssistantIdentity into assistant/ (LUM-1753.4)
These were sitting in `domains/chat/api/assistant.ts` — the
runtime identity of the assistant has nothing to do with chat.
It's a property of the assistant itself, queried by chat, the
identity tab, contacts, intelligence, onboarding, the sync
router, and several stream handlers. Misplaced since day one.

Now lives at `apps/web/src/assistant/identity.ts`. The `chat`
file is renamed only in its header comment — it now solely owns
chat-context bootstrapping (`getChatContext`, `fetchAssistantId`).
The `domains/chat/api/assistant.ts` filename stays for now to
avoid churn; can be renamed in a follow-up if it bothers anyone.

Import sites updated: identity-tab, use-assistant-identity-init,
chat-page (two lines), chat-route-content, contacts-page,
chat-utils.

Cross-domain allow-list: 119 → 118 imports. Small numeric drop —
the structural win is that `fetchAssistantIdentity` no longer
sits in another feature's folder.

Verified locally: lint, typecheck, tests all green.
